Added quadratic curve + rename to drawRoundedShape
